Small tweaks
